+++ IbFlexReader/IbFlexReader.Contracts/Enums/AssetCategory.cs
@@ -1,0 +1,7 @@
+﻿namespace IbFlexReader.Contracts.Enums
+{
+    public enum AssetCategory
+    {
+        STK, OPT, FOP, CFD, FUT
+    }
+}
