+++ IbFlexReader/IbFlexReader.Contracts/Enums/BuySell.cs
@@ -1,0 +1,7 @@
+﻿namespace IbFlexReader.Contracts.Enums
+{
+    public enum BuySell
+    {
+        BUY, SELL
+    }
+}
