+++ IbFlexReader/IbFlexReader.Contracts/Enums/CashTransactionType.cs
@@ -1,0 +1,21 @@
+﻿using EnumParser;
+
+namespace IbFlexReader.Contracts.Enums
+{
+    [EnumName]
+    public enum CashTransactionType
+    {
+        [EnumName("Other Fees")]
+        OtherFees,
+        [EnumName("Withholding Tax")]
+        WithholdingTax,
+        [EnumName("Dividends")]
+        Dividends,
+        [EnumName("Payment In Lieu Of Dividends")]
+        PaymentInLieuOfDividends,
+        [EnumName("Broker Interest Paid")]
+        BrokerInterestPaid,
+        [EnumName("Price Adjustments")]
+        PriceAdjustments
+    }
+}
