+++ IbFlexReader/IbFlexReader.Contracts/Enums/Currencies.cs
@@ -1,0 +1,15 @@
+﻿namespace IbFlexReader.Contracts.Enums
+{
+    public enum Currencies
+    {
+        EUR,
+        USD,
+        JPY,
+        CHF,
+        GBP,
+        NZD,
+        AUD,
+        CAD,
+        SEK
+    }
+}
