+++ IbFlexReader/IbFlexReader.Contracts/Enums/LongShort.cs
@@ -1,0 +1,7 @@
+﻿namespace IbFlexReader.Contracts.Enums
+{
+    public enum LongShort
+    {
+        Long, Short
+    }
+}
