+++ IbFlexReader/IbFlexReader.Contracts/Enums/OpenClose.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace IbFlexReader.Contracts.Enums
+{
+    [Flags]
+    public enum OpenClose
+    {
+        O = 1, C = 2
+    }
+}
