+++ IbFlexReader/IbFlexReader.Contracts/Enums/PutCall.cs
@@ -1,0 +1,7 @@
+﻿namespace IbFlexReader.Contracts.Enums
+{
+	public enum PutCall
+	{
+		P, C
+	}
+}
