+++ IbFlexReader/IbFlexReader.Tests/IbFlexReader.Tests.csproj
@@ -41,15 +41,6 @@
     <Reference Include="FluentAssertions, Version=5.5.3.0, Culture=neutral, PublicKeyToken=33f2691a05b67b6a, processorArchitecture=MSIL">
       <HintPath>..\packages\FluentAssertions.5.5.3\lib\net45\FluentAssertions.dll</HintPath>
     </Reference>
-    <Reference Include="IbFlexReader">
-      <HintPath>..\IbFlexReader\bin\Debug\netstandard2.0\IbFlexReader.dll</HintPath>
-    </Reference>
-    <Reference Include="IbFlexReader.Contracts">
-      <HintPath>..\IbFlexReader\bin\Debug\netstandard2.0\IbFlexReader.Contracts.dll</HintPath>
-    </Reference>
-    <Reference Include="IbFlexReader.Xml">
-      <HintPath>..\IbFlexReader\bin\Debug\netstandard2.0\IbFlexReader.Xml.dll</HintPath>
-    </Reference>
     <Reference Include="nunit.framework, Version=3.11.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
       <HintPath>..\packages\NUnit.3.11.0\lib\net45\nunit.framework.dll</HintPath>
     </Reference>
@@ -66,16 +57,35 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Helper.cs" />
+    <Compile Include="StringFactory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReaderRealTest.cs" />
-    <Compile Include="TradeTest.cs" />
-    <Compile Include="XmlTest.cs" />
+    <Compile Include="Xml\TradeTest.cs" />
+    <Compile Include="Xml\XmlTest.cs" />
     <Compile Include="XmlTradeConfirmsTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\IbFlexReader.Contracts\IbFlexReader.Contracts.csproj">
+      <Project>{21a0a357-2533-429b-ab35-a9d5497263e7}</Project>
+      <Name>IbFlexReader.Contracts</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\IbFlexReader.Utils\IbFlexReader.Utils.csproj">
+      <Project>{5401088f-eccf-4e3d-8d99-6c31d18e6571}</Project>
+      <Name>IbFlexReader.Utils</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\IbFlexReader.Xml\IbFlexReader.Xml.csproj">
+      <Project>{9a8831da-29e8-4f17-b2b5-3c4bd0ed54a5}</Project>
+      <Name>IbFlexReader.Xml</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\IbFlexReader\IbFlexReader.csproj">
+      <Project>{9e8052d9-df98-49ac-b5d2-b1f254330c8f}</Project>
+      <Name>IbFlexReader</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
