+++ IbFlexReader/IbFlexReader.Tests/ReaderRealTest.cs
@@ -1,19 +1,17 @@
 ﻿using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace IbFlexReader.Test
-{    
+namespace IbFlexReader.Tests
+{
     public class ReaderRealTest
     {
-        private static string queryId = "";
-        private static string token = "";
+        private string _queryId = string.Empty;
+        private string _token = string.Empty;
 
         [Test]
         public void Test()
         {
-            var result = new Reader().GetByApi(token, queryId);
+            var result = new Reader().GetByApi(_token, _queryId);
+            Assert.IsNotNull(result);
 
         }
     }
