+++ IbFlexReader/IbFlexReader.Xml/Contracts/QueryResponse.cs
