+++ IbFlexReader/IbFlexReader/Reader.cs
@@ -2,7 +2,6 @@
 using IbFlexReader.Xml;
 using IbFlexReader.Xml.Contracts;
 using System;
-using System.IO;
 using System.Net.Http;
 using System.Threading.Tasks;
 
@@ -10,22 +9,46 @@ namespace IbFlexReader
 {
     public class Reader
     {
-        public Contracts.FlexQueryResponse GetByString(string content, Options options = null)
+        private  IStreamBuilder sb;
+
+        
+        public Reader()
         {
-            var stream = GenerateStreamFromString(content);
+            sb = new StreamBuilder();
+        }
+
+        /// <summary>
+        ///  
+        /// </summary>
+        /// <param name="convertedXmlFile"></param>
+        /// <param name="options"></param>
+        /// <returns>An object with account activities</returns>
+        public Contracts.FlexQueryResponse GetByString(string convertedXmlFile, Options options = null)
+        {
+
+            var stream = this.sb.GenerateStreamFromString(convertedXmlFile);
+
             var result = Deserializer.Deserialize<Xml.Contracts.FlexQueryResponse, Contracts.FlexQueryResponse>(stream, out var errorObjects);
             result = result ?? new Contracts.FlexQueryResponse();
             result.Errors = errorObjects;
             Logic.ProcessStatement(result.FlexStatements?.FlexStatement, options);
+
             return result;
         }
 
-        public async Task<Contracts.FlexResult> GetByApi(string token, string queryId)
+
+
+        
+        public async Task<FlexResult> GetByApi(string token, string queryId)
         {
+            // todo: cover this code with tests
             var client = new HttpClient();
             var uri = new Uri($"https://gdcdyn.interactivebrokers.com/Universal/servlet/FlexStatementService.SendRequest?t={token}&q={queryId}&v=3");
+
             var result = await client.PostAsync(uri, null);
+
             var stream = await result.Content.ReadAsStreamAsync();
+
             var response = Deserializer.Deserialize<XmlFlexStatementResponse, FlexStatementResponse>(stream, out var errorObjects);
 
             if (response.Status != "Success")
@@ -36,6 +59,7 @@ namespace IbFlexReader
                     ErrorMessage = response.ErrorMessage
                 };
             }
+           
 
             uri = new Uri($"https://gdcdyn.interactivebrokers.com/Universal/servlet/FlexStatementService.GetStatement?q={response.ReferenceCode}&t={token}&v=3");
             result = await client.PostAsync(uri, null);
@@ -50,14 +74,6 @@ namespace IbFlexReader
             };
         }
 
-        public static Stream GenerateStreamFromString(string s)
-        {
-            var stream = new MemoryStream();
-            var writer = new StreamWriter(stream);
-            writer.Write(s);
-            writer.Flush();
-            stream.Position = 0;
-            return stream;
-        }
+
     }
 }
